Temporary fixes for RangeModel
The behavior is more or less self-consistant. What needs to be checked
is how to keep the Slider in a consistant state after creation and
initialisation, where the order of evaluation for the different
properties is not guaranteed.

Also added ColorChooser demo (with convinience SlideBoxr component) to
test Slider-SpinBox bindings.