Separate exclusive logic from layout in ButtonGroup
A new class is introduced, ExclusiveGroup. It manages a set of objects
having the "checked" property in a mutually exclusive way. To do that,
it needs to have each object presented to it through the function
ExclusiveGroup.connectCheckedChanged(). There, it will listen to the
"checkedChanged" signal for each object.

Also added JSUtils.js, containing some convenience JS functions.

Reviewed-by: Jens