Symbian: Single tap on the text cursor shouldn't open the context menu
In Avkon editors, the text context menu is shown only when the field
is long pressed or text is selected. In qt-components editors, the
context menu is currently shown also when the text field is tapped
on the text cursor (this functionality was originally implemented as
it was requested by the UX team, see QTCOMPONENTS-1081).
However now it has been decided that the context menu shouldn't be
opened anymore with a single tap on the text cursor as this has
created some usability issues; the qt-components editors should
also be in general more in line with Avkon editors and thus it's
also now possible to open the context menu with a long tap in read
only mode.

Furthermore, the "select" buttons are no longer shown in the menu
if the text field is empty. And the context menu itself is now
shown only if it contains at least one functional button. One tried
also to diminish the flicker that one can sometimes see when the
menu disappears by hiding the menu before calculating its new
position in TextTouchController's onCursorPositionChanged method.

Task-number: ou1cimx1#1001854

 Author:    Riikka Inkila <ext-riikka.j.inkila@nokia.com>

 Reviewed-by: Tuija Lindfors <ext-tuija.lindfors@nokia.com>