Symbian: Fix for certain problems caused by z-order value Infinity
Several both internal and 3rd party components have their z-order
value defined as Infinity.

This may cause problems especially when the touch context tools
and magnifier glass have to be on top of an editor which has as a
parent a popup component.

To ensure that the tools and magnifier glass are always on top,
a visibility signal check is added to the text editor components.

Task-number: ou1cimx1#1003011
Merge-request: 1358
Reviewed-by: Jaakko Helanti <ext-jaakko.helanti@nokia.com>
Reviewed-by: Mikael Fagerlund <ext-mikael.fagerlund@nokia.com>
Reviewed-by: Pauli Jarvinen <ext-pauli.p.jarvinen@nokia.com>
(cherry picked from commit d6c90cdcca0eccef94ba4d988be9dfe7055cf778)